+++ frontend/src/pages/publik/kamus/index.js
@@ -1,4 +1,5 @@
 export { default as Ejaan } from './Ejaan';
+export { default as Gramatika } from './Gramatika';
 export { default as Kamus } from './Kamus';
 export { default as KamusDetail } from './KamusDetail';
 export { default as Makna } from './Makna';